+++ db/cluster/cluster_test.go
@@ -1,12 +1,20 @@
 package cluster
 
 import (
+	"reflect"
+	"testing"
+	"time"
+)
+
+/*
+import (
 	"flag"
 	"fmt"
 	"reflect"
 	"testing"
 	"time"
 )
+
 
 var argRunClusterTests = false
 
@@ -40,6 +48,7 @@ func getOpts() Opts {
 	}
 	return res
 }
+
 func TestBasic(t *testing.T) {
 	if !argRunClusterTests {
 		t.Skip("pass cluster-tests-run to enable")
@@ -60,18 +69,6 @@ func TestBasic(t *testing.T) {
 	}
 }
 
-func assertEq(t *testing.T, want, got interface{}, labels ...string) {
-	t.Helper()
-	if !reflect.DeepEqual(want, got) {
-		t.Errorf("wanted %v, got %v, labels %v", want, got, labels)
-	}
-}
-
-func date(m, s int) time.Time {
-	return time.Date(2018, 8, 29, 23, m, s, 1, time.UTC)
-}
-
-/*
 func TestBasicRun(t *testing.T) {
 	if !argRunClusterTests {
 		t.Skip("pass cluster-tests-run to enable")
@@ -89,3 +86,14 @@ func TestBasicRun(t *testing.T) {
 	time.Sleep(10 * time.Second)
 }
 */
+
+func assertEq(t *testing.T, want, got interface{}, labels ...string) {
+	t.Helper()
+	if !reflect.DeepEqual(want, got) {
+		t.Errorf("wanted %v, got %v, labels %v", want, got, labels)
+	}
+}
+
+func date(m, s int) time.Time {
+	return time.Date(2018, 8, 29, 23, m, s, 1, time.UTC)
+}
