--- db/cluster2/driver/driver.go
@@ -1,1 +1,0 @@
-package driver
